Split behavior default method support
