+++ src/typechecker/behavior_impl_support/default_methods.rs
@@ -1,0 +1,196 @@
+use super::*;
+
+impl TypeChecker {
+    pub(in crate::typechecker) fn behavior_default_methods_for_impl(
+        &self,
+        type_name: &str,
+        behavior: &str,
+        behavior_type_args: &[AstType],
+        methods: &[Declaration],
+    ) -> Vec<DefaultBehaviorMethod> {
+        let behavior_substitutions =
+            self.behavior_type_param_substitutions(behavior, behavior_type_args);
+        self.behavior_methods_for_impl(behavior, &behavior_substitutions, &mut HashSet::new())
+            .iter()
+            .filter(|required| {
+                required.default_body.is_some()
+                    && !self.impl_methods_include_behavior_method(
+                        type_name,
+                        methods,
+                        &required.name,
+                    )
+            })
+            .filter_map(|required| {
+                let body = required.default_body.clone()?;
+                Some(DefaultBehaviorMethod {
+                    name: required.name.clone(),
+                    params: required
+                        .params
+                        .iter()
+                        .map(|param| Param {
+                            name: param.name.clone(),
+                            ty: concrete_self_ast_type(&param.ty, type_name),
+                            mutable: param.mutable,
+                            span: param.span,
+                        })
+                        .collect(),
+                    return_type: required
+                        .return_type
+                        .as_ref()
+                        .map(|ty| concrete_self_ast_type(ty, type_name)),
+                    body,
+                    span: required.span,
+                })
+            })
+            .collect()
+    }
+
+    pub(in crate::typechecker) fn seed_behavior_default_method_signature(
+        &mut self,
+        type_name: &str,
+        default: &DefaultBehaviorMethod,
+    ) {
+        let key = Self::method_key(type_name, &default.name);
+        self.methods.insert(
+            key.clone(),
+            func_info_from_behavior_method(key, &default.params, &default.return_type),
+        );
+    }
+
+    pub(in crate::typechecker) fn impl_methods_include_behavior_method(
+        &self,
+        type_name: &str,
+        methods: &[Declaration],
+        required_name: &str,
+    ) -> bool {
+        methods
+            .iter()
+            .any(|decl| matches!(decl, Declaration::Function { name, .. } if name == required_name))
+            || (self.resolver_backed_collection
+                && self
+                    .resolver_backed_method_signature(type_name, required_name)
+                    .is_some())
+    }
+
+    pub(in crate::typechecker) fn behavior_methods_with_inherited_substituted(
+        &self,
+        behavior: &str,
+        substitutions: &HashMap<String, AstType>,
+        seen: &mut HashSet<String>,
+    ) -> Vec<ast::BehaviorMethod> {
+        if !self.mark_behavior_seen(behavior, substitutions, seen) {
+            return Vec::new();
+        }
+
+        let mut methods = Vec::new();
+        if let Some(parents) = self.behavior_extends.get(behavior) {
+            for parent in parents {
+                let parent_substitutions =
+                    self.behavior_parent_type_param_substitutions(parent, substitutions);
+                methods.extend(self.behavior_methods_with_inherited_substituted(
+                    &parent.behavior,
+                    &parent_substitutions,
+                    seen,
+                ));
+            }
+        }
+        if let Some(info) = self.behaviors.get(behavior) {
+            methods.extend(
+                info.methods
+                    .iter()
+                    .map(|method| substituted_behavior_method_signature(method, substitutions)),
+            );
+        }
+        methods
+    }
+
+    pub(in crate::typechecker) fn behavior_seen_key(
+        &self,
+        behavior: &str,
+        substitutions: &HashMap<String, AstType>,
+    ) -> String {
+        let type_args = self
+            .behaviors
+            .get(behavior)
+            .map(|info| {
+                info.type_params
+                    .iter()
+                    .filter_map(|param| substitutions.get(param).cloned())
+                    .collect::<Vec<_>>()
+            })
+            .unwrap_or_default();
+        behavior_ref_display(behavior, &type_args)
+    }
+
+    pub(in crate::typechecker) fn mark_behavior_seen(
+        &self,
+        behavior: &str,
+        substitutions: &HashMap<String, AstType>,
+        seen: &mut HashSet<String>,
+    ) -> bool {
+        let behavior_seen_key = self.behavior_seen_key(behavior, substitutions);
+        seen.insert(behavior_seen_key)
+    }
+
+    pub(in crate::typechecker) fn behavior_methods_for_impl(
+        &self,
+        behavior: &str,
+        substitutions: &HashMap<String, AstType>,
+        seen: &mut HashSet<String>,
+    ) -> Vec<ast::BehaviorMethod> {
+        self.behavior_methods_with_inherited_substituted(behavior, substitutions, seen)
+    }
+
+    pub(in crate::typechecker) fn behavior_type_param_substitutions(
+        &self,
+        behavior: &str,
+        type_args: &[AstType],
+    ) -> HashMap<String, AstType> {
+        self.behaviors
+            .get(behavior)
+            .map(|info| {
+                info.type_params
+                    .iter()
+                    .cloned()
+                    .zip(type_args.iter().cloned())
+                    .collect()
+            })
+            .unwrap_or_default()
+    }
+
+    pub(in crate::typechecker) fn behavior_parent_type_param_substitutions(
+        &self,
+        parent: &BehaviorParentRef,
+        substitutions: &HashMap<String, AstType>,
+    ) -> HashMap<String, AstType> {
+        let parent_type_args: Vec<AstType> = parent
+            .type_args
+            .iter()
+            .map(|type_arg| substitute_behavior_ast_type(type_arg, substitutions))
+            .collect();
+        self.behavior_type_param_substitutions(&parent.behavior, &parent_type_args)
+    }
+
+    pub(in crate::typechecker) fn impl_ast_types_compatible(
+        &self,
+        expected: &AstType,
+        actual: &AstType,
+        self_type_name: &str,
+    ) -> bool {
+        match expected {
+            AstType::SelfType => matches!(actual, AstType::Named(name) if name == self_type_name),
+            _ => expected == actual,
+        }
+    }
+
+    pub(in crate::typechecker) fn impl_type_display(
+        &self,
+        ty: &AstType,
+        self_type_name: &str,
+    ) -> String {
+        match ty {
+            AstType::SelfType => self_type_name.to_string(),
+            _ => ty.display_name(),
+        }
+    }
+}
